+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.9
+- feat: Removed unhandled banned class names
+- feat: Simplified regex object definition in config
+
 ## 0.0.8
 - feat: Banned layers config severity